feat: Improve protected route logic and add redirect for unauthorized access
The code changes in the protected-route.tsx component improve the logic for handling protected routes. If the user is not authenticated or does not have the necessary roles (isAdmin, isTaytay, isPasig), they will be redirected to the login page. Additionally, if the user is authenticated but tries to access a path other than "/job-orders" or "/account", they will be redirected to the "/job-orders" page. These changes enhance the security and user experience of the application.