feat: Update useUser.ts to allow multiple admin emails
The code changes in useUser.ts modify the isAdmin check to allow multiple admin emails. Previously, only one email (avisha@email.com) was considered as an admin. With this change, both "avisha@email.com" and "admin@admin.com" are recognized as admin emails. This enhances the flexibility and scalability of the admin functionality in the application.