bug fix: Add discount field to JobOrderForm
The code changes in job-order-form.tsx add a new "discount" field to the JobOrderForm component. This allows users to apply a discount to the job order, which will be reflected in the calculations for total materials cost, downpayment, and grand total. This enhancement improves the flexibility and accuracy of the job order form.